Preparation code for Ticket Buy API in the future

+++ BackEnd/park_management/src/main/java/park/management/com/vn/model/request/StaffAssignmentRequest.java
@@ -15,8 +15,8 @@ public class StaffAssignmentRequest {
     private LocalDate assignedDate;
 
     @NotNull
-    private Integer staffId;
+    private Long staffId;
 
     @NotNull
-    private Integer shiftId;
+    private Long shiftId;
 }
